+++ demo/ucumDemo.js
@@ -312,7 +312,8 @@ export class UcumDemo {
    * This method is run when the Converter tab is displayed, to reset the
    * variables that track the validity of the three inputs ("from" unit code,
    * "from" value, and "to" unit code) that must be correct to attempt a
-   * conversion.
+   * conversion as well as to reset the molecular weight division to hidden
+   * status and the associated moleWeight input field to null.
    */
   showConvertTab(){
     this.convFrom_ = false ;
@@ -340,6 +341,12 @@ export class UcumDemo {
     let toNumField = document.getElementById('convertToNum');
     toNumField.value = 1;
     toNumField.classList.remove("invalid");
+
+    let mWeightDiv = document.getElementById('molecular-weight');
+    mWeightDiv.style.visibility = 'hidden';
+
+    let mWeightFld = document.getElementById('moleWeight');
+    mWeightFld.value = null ;
 
     if (window.location.hash !== '#converter' &&
         window.location.hash !== '') {
@@ -459,7 +466,7 @@ export class UcumDemo {
    *  string to be validated
    * @param tabName the name of the tab ('Converter' or 'Validator') that
    *  contains the input field
-   * @param msgElementID the ID of the web page element to received messages
+   * @param msgElementID the ID of the web page element to receive messages
    *  (success and error).
    * @returns nothing directly; return is the success/error message
    */
@@ -478,6 +485,17 @@ export class UcumDemo {
       suggsField = document.getElementById("convSuggestions");
       suggsField.innerHTML = '';
       suggsField.style.display = 'none';
+      // This is called by a change to one of the unit fields, so
+      // hide the molecular weight division, clear the moleWeight input
+      // value, and clear the value of the last result field, if any
+      let mWeightDiv = document.getElementById('molecular-weight');
+      mWeightDiv.style.visibility = 'hidden';
+      let mWeightFld = document.getElementById('moleWeight');
+      mWeightFld.value = null ;
+      if (this.lastResultFld_) {
+        let rf = document.getElementById(this.lastResultFld_);
+        rf.value = '';
+      }
     }
 
     let resFld = document.getElementById(msgElementID);
@@ -658,9 +676,15 @@ export class UcumDemo {
       this.convFromNum_ = valid ;
       resultSide = 'to';
     }
-    else { // assume convertToNum
+    else if (elementID === 'convertToNum') {
       this.convToNum_ = valid;
       resultSide = 'from';
+    }
+    else { // assume moleWeight
+      if (this.lastResultFld_ === 'convertFromNum')
+        resultSide = 'from';
+      else
+        resultSide = 'to';
     }
 
     // set or remove the indicator (currently a red border) on the invalid field
@@ -704,9 +728,12 @@ export class UcumDemo {
 
     let checkFld = document.getElementById(numField);
     let checkVal = escapeHtml(checkFld.value);
-    let resFld = numField === 'convertFromNum' ?
-                 document.getElementById('convertToNum') :
-                 document.getElementById('convertFromNum') ;
+    if (numField === 'moleWeight')
+      var resFld = document.getElementById(this.lastResultFld_) ;
+    else
+      resFld = numField === 'convertFromNum' ?
+               document.getElementById('convertToNum') :
+               document.getElementById('convertFromNum') ;
     let resultString = document.getElementById('convMsg');
     resultString.innerHTML = '';
 
@@ -801,9 +828,14 @@ export class UcumDemo {
       decDigits = Ucum.decDigits_;
       prec.value = decDigits;
     }
-    // Start out with the precision chooser line hidden
+    // Start out with the precision not displayed
     let precLine = document.getElementById('precision-line');
-    precLine.style.visibility = "hidden";
+    precLine.style.display = "none";
+
+    // Store the name of the result field so that we have it if a larger number
+    // of digits is subsequently requested AND (for the result field name)
+    // when a molecular weight is requested.
+    this.lastResultFld_ = toNumField ;
 
     let fromName = document.getElementById(fromField).value ;
     let escFromName = escapeHtml(fromName);
@@ -822,14 +854,16 @@ export class UcumDemo {
     suggsField.innerHTML = '';
     suggsField.style.display = 'none';
 
+    let weightField = document.getElementById("moleWeight");
+    let moleWeightVal = weightField.value ;
+
     let resultObj = this.utils_.convertUnitTo(fromName, fromVal, toName,
-                                                true);
+                                                true, moleWeightVal);
     if (resultObj['status'] === 'succeeded') {
       let toVal = resultObj['toVal'];
 
-      // Store the name of the result field and the full number received so
-      // that we have it if a larger number of digits is subsequently requested.
-      this.lastResultFld_ = toNumField ;
+      // Store the full result value number received so that we have it if a
+      // larger number of digits is subsequently requested.
       this.lastResult_ = toVal;
 
       // convert the value to a fixed value with the specified number of
@@ -868,10 +902,22 @@ export class UcumDemo {
         'perform the conversion.';
     }
 
-    // Else 1 or more invalid unit expressions were found (status = 'failed')
+    // Else result status was failed.  Either a molecular weight is needed
+    // for a mass<->moles conversion or 1 or more invalid unit expressions
+    // were found (status = 'failed')
     else {
       if (resultObj['msg']) {
-        resultMsg = resultObj['msg'].join('<BR>');
+        let idx = resultObj['msg'].indexOf(Ucum.needMoleWeightMsg_);
+        if (idx >= 0) {
+          this._requestMolecularWeight();
+          resultObj['msg'].splice(idx, 1);
+          if (resultObj['msg'].length > 0)
+            resultMsg = resultObj['msg'].join('<BR>');
+          else
+            resultMsg = '';
+        } else {
+          resultMsg = resultObj['msg'].join('<BR>');
+        }
       }
       // if suggestions were found, output the suggestions to the suggestions
       // field
@@ -1219,6 +1265,18 @@ export class UcumDemo {
       }
     }
   } // end _sizeNameDivs
+
+
+  /**
+   * This makes the division with the request for molecular weight
+   * visible.
+   *
+   * @private
+   */
+  _requestMolecularWeight() {
+    let weightDiv = document.getElementById('molecular-weight');
+    weightDiv.style.visibility = 'visible';
+  }
 
 } // end class UcumDemo