+++ CHANGELOG.md
@@ -4,6 +4,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## [2.6.1] - 2019-03-22
+### Added
+- Added mass<-> moles conversion to the UCUM Demo page, with a request to the
+user to supply a molecular weight when such a conversion is requested.
+
 ## [2.6.0] - 2019-03-08
 ### Added
 - Added capability to existing convertUnitTo library function to convert mass